feat: emit habit created event when a new habit is created
To emit events and test that they have been emitted using Spring
Modulith, we need Spring Data JDBC. Alternatively, Spring Modulith
supports Spring Data JPA and Spring Data MongoDB.

The problem is that all supported persistence technologies are
blocking and not reactive. Since our application is reactive, it
requires a reactive database driver as well, to persist aggregates in
a reactive way.

Another issue is that due to both Spring Data JDBC and Spring
Data RD2BC on the classpath, we have two `TransactionManager`
beans. A `PlatformTransactionManager` and a
`ReactiveTransactionManager`. Spring Modulith uses @Transactional
without qualifier, which means that executing annotated methods will
throw a `NoUniqueBeanDefinitionException`.

To workaround this issue, we define a `PlatformTransactionManager`
manually.